Added checks for existing files, skips if file exists

+++ AutoFileCreation.py
@@ -1,6 +1,7 @@
 import shutil
 import os
 from dotenv import load_dotenv
+from pathlib import Path
 
 # Load environment variables from .env file
 load_dotenv()
@@ -19,13 +20,12 @@ new_model_item_name_camel_case = new_value[0].lower() + new_value[1:]
 example_item_name_camel_case = example_item_name[0].lower() + example_item_name[1:]
 
 
-
 def main():
     # ------ Start: Put each set of file names and locations here
 
     # IStorage
     i_storage_file_name = 'IStorageBroker._s.cs'
-    location = f'{destination}{app_name}.Api\\Brokers\\Storages\\'      
+    location = f'{destination}{app_name}.Api\\Brokers\\Storages\\'
     copy_example_to_new_file(i_storage_file_name, location)
 
     # Storage
@@ -39,26 +39,30 @@ def main():
 def copy_example_to_new_file(fill_in_file_name, file_path):
     # Copy to empty file
     new_file_name = fill_in_file_name.replace('_', new_value)
-    
-    if (is_mac == 'true'):
-        file_path.replace('\\', '/')
-    
+    if is_mac == 'true':
+        file_path = file_path.replace('\\', '/')
+
     example_item_file_name = fill_in_file_name.replace('_', example_item_name)
     new_file_path = f'{file_path}{new_file_name}'
-        
-    shutil.copyfile(f'{file_path}{example_item_file_name}', new_file_path)
 
-    # Replace content
-    with open(new_file_path, 'r') as file: content = file.read()
+    copied_file_path = Path(new_file_path)
+    if copied_file_path.is_file():
+        print(new_file_path + ' already exists, skipping')
+    else:
+        shutil.copyfile(f'{file_path}{example_item_file_name}', new_file_path)
 
-    # Replace all instances of the title case variable
-    content = content.replace(example_item_name, new_value)
+        # Replace content
+        with open(new_file_path, 'r') as file: content = file.read()
 
-    # Replace all instances of the camelCase variable
-    content = content.replace(example_item_name_camel_case, new_model_item_name_camel_case)
+        # Replace all instances of the title case variable
+        content = content.replace(example_item_name, new_value)
 
-    # Write the modified content back to the file
-    with open(new_file_path, 'w') as file: file.write(content)
+        # Replace all instances of the camelCase variable
+        content = content.replace(example_item_name_camel_case, new_model_item_name_camel_case)
+
+        # Write the modified content back to the file
+        with open(new_file_path, 'w') as file: file.write(content)
+        print('Created '+new_file_path)
 
 
 if __name__ == "__main__":
